Renamed tailgating to tailgating_detection
